# crates/sdkwork-intelligence-agents-service/specs/openapi/agents-backend-api.openapi.yaml
openapi: 3.1.2
info:
  title: SDKWork Agent Business Backend API
  version: 5.0.0
  description: Backend-facing managed agent execution APIs for SDKWork operators.
...
servers:
- url: http://localhost:8080
...
paths:
  /backend/v3/api/ai/agents:
    get:
      tags:
      - ai
      summary: List managed agents for backend administration
      operationId: agents.list
      security:
      - AuthToken: []
        AccessToken: []
      x-sdkwork-domain: ai
      x-sdkwork-resource: agents
      x-sdkwork-permission: agent.business.list
      x-sdkwork-tenant-scope: tenant
      x-sdkwork-audit-event: agent.business.listed
      parameters:
      - $ref: '#/components/parameters/IncludeDeleted'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      - $ref: '#/components/parameters/SearchQ'
      responses:
        '200':
          description: Managed agent list response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      tags:
      - ai
      summary: Create a managed agent
      description: 'Any authenticated user can create a managed agent. The agent is
        created

        in `draft` status and must be activated (draft to active) by a user with

        `ai.agents.manage` permission, providing the review workflow.

        '
      operationId: agents.create
      security:
      - AuthToken: []
        AccessToken: []
      x-sdkwork-domain: ai
      x-sdkwork-resource: agents
      x-sdkwork-permission: agent.business.create
      x-sdkwork-permission-scope: authenticated
      x-sdkwork-tenant-scope: tenant
      x-sdkwork-audit-event: agent.business.created
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Created managed agent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '403':
          description: Permission denied
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: https://sdkwork.dev/problems/permission_required
                title: permission_required
                status: 403
                detail: agent management denied
                code: 40301
                errorCategory: permission
                retryable: false
                traceId: 00000000-0000-0000-0000-000000000000
        '409':
          description: Resource conflict (duplicate agent_id or code within tenant
            scope)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: https://sdkwork.dev/problems/conflict
                title: conflict
                status: 409
                detail: agent already exists
                code: 40901
                errorCategory: business
                retryable: false
                traceId: 00000000-0000-0000-0000-000000000000
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    AuthToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AccessToken:
      type: apiKey
      in: header
      name: Access-Token
  parameters:
    IncludeDeleted:
      name: include_deleted
      in: query
      required: false
      schema:
        type: boolean
        default: false
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
...
    SearchQ:
      name: q
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  responses:
    Problem:
      description: RFC 9457 problem detail response
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            ValidationError:
              summary: Validation failure
              value:
                type: https://sdkwork.dev/problems/validation_error
                title: validation_error
                status: 400
                detail: requestedAt must be RFC3339 date-time
                code: 40001
                errorCategory: validation
                retryable: false
                traceId: 00000000-0000-0000-0000-000000000000
            VersionConflict:
              summary: Optimistic concurrency conflict
              value:
                type: https://sdkwork.dev/problems/version_conflict
                title: version_conflict
                status: 409
                detail: 'agent version mismatch: expected=2, actual=3'
                code: 40901
                errorCategory: concurrency
                retryable: true
                traceId: 00000000-0000-0000-0000-000000000000
            PermissionRequired:
              summary: Permission denied
              value:
                type: https://sdkwork.dev/problems/permission_required
                title: permission_required
                status: 403
                detail: agent management denied
                code: 40301
                errorCategory: permission
                retryable: false
                traceId: 00000000-0000-0000-0000-000000000000
            NotFound:
              summary: Resource not found
              value:
                type: https://sdkwork.dev/problems/not_found
                title: not_found
                status: 404
                detail: agent not found
                code: 40401
                errorCategory: resource
                retryable: false
                traceId: 00000000-0000-0000-0000-000000000000
  schemas:
...
    SdkWorkApiResponse:
      type: object
      additionalProperties: false
      required:
      - code
      - data
      - traceId
      properties:
        code:
          type: integer
          format: int32
          description: Numeric success result code. MUST be 0 on HTTP 2xx JSON bodies.
            See API_SPEC.md 搂15.3.
          enum:
          - 0
          default: 0
          minimum: 0
          maximum: 0
        data:
          description: Operation-specific payload. Typed per operation through allOf
            or explicit schema refs.
        traceId:
          type: string
          format: uuid
          description: Server-owned request correlation id. Clients MUST NOT supply
            this value.
...
    SdkWorkPageData:
      type: object
      additionalProperties: false
      required:
      - items
      - pageInfo
      properties:
        items:
          type: array
          items: {}
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
    Int64String:
      type: string
      pattern: ^[0-9]+$
...
    AgentVisibility:
      type: string
      enum:
      - private
      - organization
      - tenant
      - public
...
    AgentImplementationKind:
      type: string
      enum:
      - manifest-only
      - typed-local-provider
      - process-adapter
      - protocol-adapter
...
    AgentRecord:
      type: object
      additionalProperties: false
      required:
      - id
      - agentId
      - tenantId
      - organizationId
      - ownerUserId
      - code
      - displayName
      - manifest
      - implementationType
      - status
      - visibility
      - tags
      - version
      - createdAt
      - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/Int64String'
        agentId:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^agent\.[a-z0-9_-]+(\.[a-z0-9_-]+)*$
        tenantId:
          $ref: '#/components/schemas/Int64String'
        organizationId:
          $ref: '#/components/schemas/Int64String'
        ownerUserId:
          $ref: '#/components/schemas/Int64String'
        code:
          type: string
          minLength: 1
          maxLength: 128
        displayName:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type:
          - string
          - 'null'
          maxLength: 4096
        manifest:
          type: object
          additionalProperties: true
        defaultCodeTaskIntent:
          type:
          - object
          - 'null'
          additionalProperties: true
        managementProfile:
          oneOf:
          - $ref: '#/components/schemas/AgentManagementProfile'
          - type: 'null'
        implementationProviderId:
          type:
          - string
          - 'null'
          minLength: 1
          maxLength: 128
          pattern: ^provider\.[a-z0-9_-]+(\.[a-z0-9_-]+)*$
        implementationKind:
          oneOf:
          - $ref: '#/components/schemas/AgentImplementationKind'
          - type: 'null'
        implementationType:
          $ref: '#/components/schemas/AgentImplementationType'
        status:
          $ref: '#/components/schemas/AgentStatus'
        visibility:
          $ref: '#/components/schemas/AgentVisibility'
        tags:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 64
          maxItems: 64
        version:
          $ref: '#/components/schemas/Int64String'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type:
          - string
          - 'null'
          format: date-time
    AgentManagementProfile:
      type: object
      additionalProperties: false
      properties:
        author:
          type: string
          minLength: 1
          maxLength: 128
        avatar:
          type: string
          minLength: 1
          maxLength: 512
        categoryId:
          type: string
          minLength: 1
          maxLength: 64
        color:
          type: string
          minLength: 1
          maxLength: 32
        debugMode:
          type: boolean
        iconName:
          type: string
          minLength: 1
          maxLength: 64
        jsonMode:
          type: boolean
        knowledgeBaseIds:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^knowledge\.base\.[a-z0-9_-]+(\.[a-z0-9_-]+)*$
          maxItems: 128
        memoryEnabled:
          type: boolean
        model:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^model\.[a-z0-9_-]+(\.[a-z0-9_-]+)*$
        skillIds:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^skill\.[a-z0-9_-]+(\.[a-z0-9_-]+)*$
          maxItems: 128
        suggestedPrompts:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 256
          maxItems: 12
        systemPrompt:
          type: string
          minLength: 1
          maxLength: 32768
        temperature:
          type: number
          minimum: 0
          maximum: 2
        toolIds:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^tool\.[a-z0-9_-]+(\.[a-z0-9_-]+)*$
          maxItems: 128
        type:
          type: string
          enum:
          - normal
          - independent
        users:
          type: string
          minLength: 1
          maxLength: 128
        voiceIds:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^voice\.[a-z0-9_-]+(\.[a-z0-9_-]+)*$
          maxItems: 16
        welcomeMessage:
          type: string
          minLength: 1
          maxLength: 4096
    AgentResponse:
      description: Single managed agent response following SdkWorkApiResponse envelope.
      allOf:
      - $ref: '#/components/schemas/SdkWorkApiResponse'
      - type: object
        required:
        - data
        properties:
          data:
            allOf:
            - $ref: '#/components/schemas/SdkWorkResourceData'
            - type: object
              required:
              - item
              properties:
                item:
                  $ref: '#/components/schemas/AgentRecord'
    AgentListResponse:
      description: Paginated managed agent list response following SdkWorkApiResponse
        envelope.
      allOf:
      - $ref: '#/components/schemas/SdkWorkApiResponse'
      - type: object
        required:
        - data
        properties:
          data:
            allOf:
            - $ref: '#/components/schemas/SdkWorkPageData'
            - type: object
              required:
              - items
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/AgentRecord'
...
    CreateAgentRequest:
      type: object
      additionalProperties: false
      required:
      - agentId
      - code
      - displayName
      - manifest
      - visibility
      - requestedAt
      properties:
        agentId:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^agent\.[a-z0-9_-]+(\.[a-z0-9_-]+)*$
        code:
          type: string
          minLength: 1
          maxLength: 128
        displayName:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type:
          - string
          - 'null'
          maxLength: 4096
        manifest:
          type: object
          additionalProperties: true
        defaultCodeTaskIntent:
          type:
          - object
          - 'null'
          additionalProperties: true
        managementProfile:
          oneOf:
          - $ref: '#/components/schemas/AgentManagementProfile'
          - type: 'null'
        implementationProviderId:
          type:
          - string
          - 'null'
          minLength: 1
          maxLength: 128
          pattern: ^provider\.[a-z0-9_-]+(\.[a-z0-9_-]+)*$
        implementationKind:
          oneOf:
          - $ref: '#/components/schemas/AgentImplementationKind'
          - type: 'null'
        implementationType:
          $ref: '#/components/schemas/AgentImplementationType'
        visibility:
          $ref: '#/components/schemas/AgentVisibility'
        tags:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 64
          maxItems: 64
        requestedAt:
          type: string
          format: date-time
...
    ProblemDetail:
      type: object
      additionalProperties: true
      required:
      - type
      - title
      - status
      - code
      - traceId
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        code:
          type: integer
          format: int32
          description: Numeric error result code. MUST be non-zero. See API_SPEC.md
            搂15.3.
          minimum: 40001
          maximum: 79999
        traceId:
          type: string
          format: uuid
          description: Server-owned request correlation id. Same semantics as SdkWorkApiResponse.traceId.
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'